requirements.txt
grpcio
grpcio-tools
pyyaml
codecov
pandas
circus
celery
python-socketio
pymongo
redis
flask
flask-cors
gunicorn
